res made delayed swapping runtime-switchable. Configuration setting   is "Video.OpenGL.DelaySwap". res enabled VBO by default. New VBO management code is worth some   more wide-spread testing. res fixed a missing "layer" attribute in water.xml. res fixed the "world2camera transform" SV not being available to   shaders when it should be. Fixes crash in waterdemo. res ported Anders Reggestad's last fix to xmlread. Fixes an infinite   loop. res changed the XML plugins to strip whitespaces before <![CDATA[   elements, fixes some shader parsing issues.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20759 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/engine/renderloop/stdsteps/generic.h
@@ -99,7 +99,7 @@ private:
   csDirtyAccessArray<csRenderMesh*> visible_meshes;
   csDirtyAccessArray<iMeshWrapper*> imeshes_scratch;
   csDirtyAccessArray<iShaderVariableContext*> mesh_svc;
-  int visible_meshes_index;	// First free index in the visible meshes.
+  size_t visible_meshes_index;	// First free index in the visible meshes.
 
   static csStringID o2c_matrix_name;
   static csStringID o2c_vector_name;
